refactor: decompose large source and test files
