+++ Cotabby/Models/Suggestion/SuggestionRequest.swift
@@ -1,0 +1,130 @@
+import Foundation
+
+/// Immutable request values shared across Apple Intelligence, llama.cpp, and endpoint backends.
+/// Engines receive this snapshot instead of consulting mutable settings or live editor state.
+
+/// Distinguishes a normal forward continuation from a typo-correction reply.
+///
+/// `.correction(typoWord:)` carries the exact word the correction was offered for, so the acceptance
+/// path can confirm the live field still ends with that word before deleting it (guarding against a
+/// keystroke that slipped in between), then replace it with the corrected word. This lets a
+/// correction reuse the same session/overlay machinery as a plain continuation while committing as a
+/// whole-word replacement.
+enum SuggestionKind: Equatable, Sendable {
+    case continuation
+    case correction(typoWord: String)
+
+    var isCorrection: Bool {
+        if case .correction = self {
+            return true
+        }
+        return false
+    }
+}
+
+/// One generation request sent from the coordinator into the suggestion engine.
+struct SuggestionRequest: Equatable, Sendable {
+    let context: FocusedInputContext
+    /// The truncated text immediately before the caret.
+    /// This stays backend-agnostic and gives every engine access to the same local writing context
+    /// even if they render prompts differently.
+    let prefixText: String
+    /// The canonical prompt payload for prompt-oriented backends such as the local llama runtime.
+    /// Engines that prefer a separate instructions channel can derive their own request text from
+    /// `prefixText` and the other shared fields instead of consuming this string directly.
+    let prompt: String
+    let generation: UInt64
+    let maxPredictionTokens: Int
+    let temperature: Double
+    let topK: Int
+    let topP: Double
+    let minP: Double
+    let repetitionPenalty: Double
+    /// Optional deterministic sampler seed. `nil` preserves production randomness; tests and
+    /// microbenches can set this so cached and uncached runtime paths are directly comparable.
+    let randomSeed: UInt32?
+    let maxSuffixCharacters: Int
+    /// Explicit length guidance stays separate from user style preferences so prompt builders can
+    /// order and phrase them differently per backend.
+    let completionLengthInstruction: String
+    /// Optional user-provided profile context. We keep this separate from base product behavior so
+    /// future settings/personalization work can evolve independently from prompt safety rules.
+    let userName: String?
+    /// User-authored style rules rendered as additional prompt directives, subordinate to the base
+    /// autocomplete/safety rules. Empty when the user has none.
+    let customRules: [String]
+    /// User-authored free-form context (glossary, jargon, style notes) injected verbatim into the
+    /// prompt. Already trimmed and length-capped upstream so renderers can treat it as a ready-to-use
+    /// string. `nil` when the user has not set it, distinguishing it from an empty-but-set value so
+    /// renderers can skip the heading entirely.
+    let extendedContext: String?
+    /// Pre-rendered language hint built from the user's declared languages (e.g. "The user usually
+    /// writes in German and English…"). `nil` when none are declared. Deliberately a hint, not an
+    /// override: it tells the model to match the surrounding text and only fall back to the declared
+    /// languages when that text is ambiguous, which protects code-switching.
+    let languageInstruction: String?
+    /// Ephemeral clipboard context captured only when the user has enabled clipboard prompting.
+    let clipboardContext: String?
+    /// Ephemeral screen context summary injected only when available for the active text field.
+    let visualContextSummary: String?
+    /// The composed writing-surface description (app class, window title, domain, placeholder),
+    /// nil when the user disabled surface context or the surface class suppresses it. The llama
+    /// prompt has already folded it in; this field exists so the Foundation Models renderer can
+    /// state the same sanitized facts in its own prompt shape.
+    let surfaceContext: SurfaceContext?
+    /// When enabled, the normalizer keeps multiple lines instead of truncating to the first line.
+    let isMultiLineEnabled: Bool
+    /// Correlation ID stamped onto every log line touching this request — coordinator state
+    /// transitions, router selection, engine generation, LLM I/O capture, insertion. Generated by
+    /// `RequestID.generate()` in `SuggestionRequestFactory`. Defaulted in the init so test fixtures
+    /// that build requests directly do not need to change.
+    let requestID: String
+
+    init(
+        context: FocusedInputContext,
+        prefixText: String,
+        prompt: String,
+        generation: UInt64,
+        maxPredictionTokens: Int,
+        temperature: Double,
+        topK: Int,
+        topP: Double,
+        minP: Double,
+        repetitionPenalty: Double,
+        randomSeed: UInt32?,
+        maxSuffixCharacters: Int,
+        completionLengthInstruction: String,
+        userName: String?,
+        customRules: [String],
+        extendedContext: String? = nil,
+        languageInstruction: String?,
+        clipboardContext: String?,
+        visualContextSummary: String?,
+        surfaceContext: SurfaceContext? = nil,
+        isMultiLineEnabled: Bool,
+        requestID: String = "req_unknown"
+    ) {
+        self.context = context
+        self.prefixText = prefixText
+        self.prompt = prompt
+        self.generation = generation
+        self.maxPredictionTokens = maxPredictionTokens
+        self.temperature = temperature
+        self.topK = topK
+        self.topP = topP
+        self.minP = minP
+        self.repetitionPenalty = repetitionPenalty
+        self.randomSeed = randomSeed
+        self.maxSuffixCharacters = maxSuffixCharacters
+        self.completionLengthInstruction = completionLengthInstruction
+        self.userName = userName
+        self.customRules = customRules
+        self.extendedContext = extendedContext
+        self.languageInstruction = languageInstruction
+        self.clipboardContext = clipboardContext
+        self.visualContextSummary = visualContextSummary
+        self.surfaceContext = surfaceContext
+        self.isMultiLineEnabled = isMultiLineEnabled
+        self.requestID = requestID
+    }
+}
